reddit/requirements.txt
praw
kafka-python